requirements_versions.txt
setuptools==69.5.1  # temp fix for compatibility with some old packages
GitPython==3.1.32
Pillow==9.5.0
accelerate==0.31.0
blendmodes==2022
clean-fid==0.1.35
diskcache==5.6.3
einops==0.4.1
facexlib==0.3.0
fastapi==0.104.1
gradio==4.40.0
httpcore==0.15
inflection==0.5.1
jsonmerge==1.8.0
kornia==0.6.7
lark==1.1.2
numpy==1.26.2
omegaconf==2.2.3
open-clip-torch==2.20.0
piexif==1.1.3
protobuf==3.20.0
psutil==5.9.5
pytorch_lightning==1.9.4
resize-right==0.0.2
safetensors==0.4.2
scikit-image==0.21.0
spandrel==0.3.4
spandrel-extra-arches==0.1.1
tomesd==0.1.3
torch==2.3.1+cu121
torchdiffeq==0.2.3
torchsde==0.2.6
transformers==4.46.1
httpx==0.24.1
pillow-avif-plugin==1.4.3
diffusers==0.32.2
gradio_rangeslider==0.0.6
gradio_imageslider==0.0.20
loadimg==0.1.2
tqdm==4.66.1
peft==0.14.0
pydantic==2.8.2
huggingface-hub==0.26.2
controlnet_aux-0.0.9-py3-none-any.whl
timm==0.9.2
onnx
onnxruntime-gpu
bitsandbytes==0.45.3
ultralytics==8.3.75
rich==13.0.0
xformers==0.0.27
aliyun_python_sdk_alimt==3.2.0
aliyun_python_sdk_core==2.13.10
boto3==1.40.6
botocore==1.40.6
cryptography==45.0.6
dill==0.4.0
distro==1.9.0
filelock==3.18.0
fonttools==4.59.0
jiter==0.10.0
jmespath==0.10.0
multiprocess==0.70.18
openai==1.99.6
pathos==0.3.4
pox==0.3.6
ppft==1.7.7
PyExecJS==1.5.1
requests==2.31.0
s3transfer==0.13.1
sentencepiece==0.2.0
typing_extensions==4.14.1
ultralytics_thop==2.0.15
urllib3==2.5.0
